requirements.txt
apache-airflow==2.5.0
pandas==1.3.5
requests==2.26.0